Refacotr: [#46] MyTicketPage Presenter/Container 컴포넌트 구조 제거
